Move PropertyValueParser and value enums to PropertyType because that's the only place where it is referenced
